requirements.txt
streamlit
transformers
langchain
langchain_community
python-dotenv
gTTS
torch
